+++ lib/client/frame.h
@@ -111,7 +111,7 @@ namespace server
          * @param toNode the receiving node.
          * @return the time in milliseconds used to compress images.
          */
-        float transmit( net::NodePtr toNode );
+        int64_t transmit( net::NodePtr toNode );
 
         /** 
          * Test the readiness of the frame.
